Protecting Video Routes

+++ controllers/videoController.js
@@ -44,10 +44,12 @@ export const postUpload = async(req, res) => {
         const newVideo = await Video.create({
             fileUrl: path.replace(/\\/g, "/"),
             title,
-            description
+            description,
+            creator: req.user.id 
         });
-        console.log(newVideo);
-        res.redirect(routes.home);
+        req.user.videos.push(newVideo.id);
+        req.user.save();
+        res.redirect(routes.videosDetail(newVideo.id));
     } catch(error){
         console.log(error);
     }
@@ -60,9 +62,10 @@ export const videosDetail = async(req, res) => {
         params: {id}
     } = req;
     try{
-      const video = await Video.findById(id);
-      res.render("videosDetail", { pageTitle: video.title, video });
+        const video = await Video.findById(id).populate("creator");
+        res.render("videosDetail", { pageTitle: video.title, video });
     } catch(error){
+        console.log(error);
         res.redirect(routes.home);
     }
 };
@@ -74,7 +77,11 @@ export const getEditVideo = async (req, res) => {
     } = req;
     try{
         const video = await Video.findById(id);
-        res.render("editVideo", {pageTitle: `Edit ${video.title}`, video});
+        if(`${video.creator}` !== `${req.user.id}`){
+            throw Error();        
+        } else{
+            res.render("editVideo", {pageTitle: `Edit ${video.title}`, video});   
+        }
     }catch(error){
         res.redirect(routes.home);
     }
@@ -100,7 +107,12 @@ export const deleteVideo = async(req, res) =>{
         params: {id}
     } = req;
     try{
-        await Video.findByIdAndRemove({ _id: id });
+        const video = await Video.findById(id);
+        if(`${video.creator}` !== `${req.user.id}`){
+            throw Error();        
+        } else{
+            await Video.findByIdAndRemove({ _id: id });
+        }
     } catch(error){
         console.log(error);
     }
